+++ src/test/java/picocli/CommandLineTest.java
@@ -151,7 +151,7 @@ public class CommandLineTest {
     }
     @Test
     public void testVersion() {
-        assertEquals("3.0.3-SNAPSHOT", CommandLine.VERSION);
+        assertEquals("3.1.0", CommandLine.VERSION);
     }
     @Test
     public void testArrayPositionalParametersAreReplacedNotAppendedTo() {