+++ RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 # picocli Release Notes
 
-# <a name="3.1.0"></a> Picocli 3.1.0 (UNRELEASED)
+# <a name="3.1.0"></a> Picocli 3.1.0
 The picocli community is pleased to announce picocli 3.1.0.
 
 This release contains bugfixes and support for command aliases.